third_party/blink/web_tests/platform/generic/css3/flexbox/flex-property-parsing-expected.txt

Tests setting the flex propery.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS flexitem.style.flex is ""
PASS getComputedStyle(flexitem).flex is "0 1 auto"
PASS flexitem.style.flex is ""
PASS getComputedStyle(flexitem).flex is "0 1 auto"
PASS flexitem.style.flex is ""
PASS getComputedStyle(flexitem).flex is "0 1 auto"
PASS flexitem.style.flex is "2 1 0%"
PASS getComputedStyle(flexitem).flex is "2 1 0%"
PASS flexitem.style.flex is "0 1 0%"
PASS getComputedStyle(flexitem).flex is "0 1 0%"
PASS flexitem.style.flex is "1.5 1 0%"
PASS getComputedStyle(flexitem).flex is "1.5 1 0%"
PASS flexitem.style.flex is "1 1 auto"
PASS getComputedStyle(flexitem).flex is "1 1 auto"
PASS flexitem.style.flex is "1 1 1px"
PASS getComputedStyle(flexitem).flex is "1 1 1px"
PASS flexitem.style.flex is "1 1 2em"
PASS getComputedStyle(flexitem).flex is "1 1 32px"
PASS flexitem.style.flex is "1 1 0px"
PASS getComputedStyle(flexitem).flex is "1 1 0px"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 0%"
PASS getComputedStyle(flexitem).flex is "0 0 0%"
PASS flexitem.style.flex is "0 1 0%"
PASS getComputedStyle(flexitem).flex is "0 1 0%"
PASS flexitem.style.flex is "1 0 0%"
PASS getComputedStyle(flexitem).flex is "1 0 0%"
PASS flexitem.style.flex is "2 1 auto"
PASS getComputedStyle(flexitem).flex is "2 1 auto"
PASS flexitem.style.flex is "3 1 4px"
PASS getComputedStyle(flexitem).flex is "3 1 4px"
PASS flexitem.style.flex is "5.25 1 auto"
PASS getComputedStyle(flexitem).flex is "5.25 1 auto"
PASS flexitem.style.flex is "4 1 6em"
PASS getComputedStyle(flexitem).flex is "4 1 96px"
PASS flexitem.style.flex is "4 1 0px"
PASS getComputedStyle(flexitem).flex is "4 1 0px"
PASS flexitem.style.flex is "0 1 0px"
PASS getComputedStyle(flexitem).flex is "0 1 0px"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "1 2 0px"
PASS getComputedStyle(flexitem).flex is "1 2 0px"
PASS flexitem.style.flex is "0 0 0px"
PASS getComputedStyle(flexitem).flex is "0 0 0px"
PASS flexitem.style.flex is "1 2 auto"
PASS getComputedStyle(flexitem).flex is "1 2 auto"
PASS flexitem.style.flex is "1.75 2 3px"
PASS getComputedStyle(flexitem).flex is "1.75 2 3px"
PASS flexitem.style.flex is "1.75 2 3px"
PASS getComputedStyle(flexitem).flex is "1.75 2 3px"
PASS flexitem.style.flex is "1.75 2 3px"
PASS getComputedStyle(flexitem).flex is "1.75 2 3px"
PASS flexitem.style.flex is "1 2 3px"
PASS getComputedStyle(flexitem).flex is "1 2 3px"
PASS flexitem.style.flex is "4 3 calc(20px + 40%)"
PASS getComputedStyle(flexitem).flex is "4 3 calc(40% + 20px)"
PASS flexitem.style.flex is "1 2 calc(10px + 50%)"
PASS getComputedStyle(flexitem).flex is "1 2 calc(50% + 10px)"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 1 auto"
PASS getComputedStyle(flexitem).flex is "0 1 auto"
PASS flexitem.style.flex is "3 1 auto"
PASS getComputedStyle(flexitem).flex is "3 1 auto"
PASS flexitem.style.flex is "0 1 0px"
PASS getComputedStyle(flexitem).flex is "0 1 0px"
PASS flexitem.style.flex is "0 1 0px"
PASS getComputedStyle(flexitem).flex is "0 1 0px"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS flexitem.style.flex is "initial"
PASS getComputedStyle(flexitem).flex is "0 1 auto"
PASS flexitem.style.flex is "0 0 auto"
PASS getComputedStyle(flexitem).flex is "0 0 auto"
PASS successfullyParsed is true

TEST COMPLETE
